pyproject.toml
[project]
name = "vast-s3-templates"
version = "0.1.0"
description = "Vast S3 Templates"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.35.53",
    "botocore==1.35.53",
    "icecream==2.1.7",
    "pandas==2.2.3",
    "python-dotenv==1.1.0",
    "s3-client==0.1.4",
]
